Fix (benign) diagnostic warnings
We were getting an RS1010 in the TeamCity build. This is harmless - it's a diagnostic rule relevant for analyzers that register multiple code fixes, but we register only one. But I want a clean build, so I've changed it anyway.

Also, skipped warning 2008 on the Vsix project, because TeamCity is raising it. (It's a complaint about there not being any source files. Oddly, VSIX projects turn out to be a special kind of C# project, and TeamCity gets upset because they have no source code.)

Related issue: #2